Started and made a possible all right amount of progress
Need to figure out exact hashing for keys don't know how this works with keys that are strings currently looking into that. Also need to remap the location of keys in map when it is resized. After that I need to figure out how to return a refrence for the [] operator. I also thought of making an additional function similar to get that returns null opt if a key is not in the map but otherwise returns the index. This would allow for checks at the head of functions that require a key to be in the map and also gets the index all at the same time.