Added emptySinceStart for buckets and getIndex + hashingfor table
Added a string multiplicative hash as found in the book as the current hashing function for the string keys. Will ask after lecture tommorow if a specific hash is supposed to be used that I am unaware of. Added emptySinceStart so that I can tell if the bucket being accessed was ever modified so that I can tell wheter or not to stop my probing efforts. If a bucket would have been the desired spot for a value but the value is not present then there is no reason to continue probing. I also added getIndex which I am using with comparison to nullopt to decide if a key is in the hash table at all. This allows me to in one line assign the table index and see if the value was in the table. This removes redundant probing searches that I had in methods like remove and get.