Change probe offset setup and fixing looping
Had a lot of for loops going one less than needed because of adding -1 to the end of the comparison. Changed how probeoffsets were setup as I still need the old offsets while I am copying over the old keys to the new data table. Currently running into the table not finding the proper value from an index search when printing out after expanding table. Looking into why this is.