Fixed hashing problem and not correctly blocking exit on resize
Previously hash was not taking into consideration that I was working with two tables when resizing so it was not hashing correctly. Also was not blocking a potential key find if the bucket was set to empty in a couple of spots. Still need to figure out the [] operator and comment the code but otherwise seems to be testing fine. I also need to ask about what kind of hashing should be done and potentially if mine appears to be doing all right.