Removed srand and fixed probe offsets. Edited hash still need to talk to prof
Realized I was not using the right values in my probe offsets array so I modified it to be 1 -> N-1 with the vector capacity being N. Also removed srand as it was mentioned in class to be unnecessary. I still need to ask what kind of hash function should be used on the key. Because of how short the strings are I am often getting in order hash table inserts.