Added getValRef functinon and finished [] operator
Changed [] operator to return size_t ref instead of int. Do not currently know if that was the right move but do not know why that would be the return when the value is a size_t not an int. need to still comment code and ensure that is fine. Also need to make bucket print out function.